remove old telemetry code
