+++ dot/node.go
@@ -18,7 +18,6 @@ package dot
 
 import (
 	"fmt"
-	"github.com/ChainSafe/gossamer/dot/tel2"
 	"net/http"
 	"os"
 	"os/signal"
@@ -353,25 +352,12 @@ func NewNode(cfg *Config, ks *keystore.GlobalKeystore, stopFunc func()) (*Node, 
 	}
 
 	telemetry.GetInstance().AddConnections(gd.TelemetryEndpoints)
-	tel2.GetTelInstance().AddConnections(gd.TelemetryEndpoints)
-	data := &telemetry.ConnectionData{
-		Authority:     cfg.Core.GrandpaAuthority,
-		Chain:         sysSrvc.ChainName(),
-		GenesisHash:   stateSrvc.Block.GenesisHash().String(),
-		SystemName:    sysSrvc.SystemName(),
-		NodeName:      cfg.Global.Name,
-		SystemVersion: sysSrvc.SystemVersion(),
-		NetworkID:     networkSrvc.NetworkState().PeerID,
-		StartTime:     strconv.FormatInt(time.Now().UnixNano(), 10),
-	}
-	telemetry.GetInstance().SendConnection(data)
-//todo ed, add test here
-	tel2.GetTelInstance().SendMessage(tel2.NewTelemetryMessage(tel2.NewKeyValue("authority", cfg.Core.GrandpaAuthority),
-		tel2.NewKeyValue("chain", sysSrvc.ChainName()), tel2.NewKeyValue("genesis_hash", stateSrvc.Block.GenesisHash().String()),
-		tel2.NewKeyValue("implementation", sysSrvc.SystemName()), tel2.NewKeyValue("msg", "system.connected"),
-		tel2.NewKeyValue("name", "EdTestNode"), tel2.NewKeyValue("network_id", networkSrvc.NetworkState().PeerID),
-		tel2.NewKeyValue("startup_time", strconv.FormatInt(time.Now().UnixNano(), 10)), tel2.NewKeyValue("version", sysSrvc.SystemVersion())))
 
+	telemetry.GetInstance().SendMessage(telemetry.NewTelemetryMessage(telemetry.NewKeyValue("authority", cfg.Core.GrandpaAuthority),
+		telemetry.NewKeyValue("chain", sysSrvc.ChainName()), telemetry.NewKeyValue("genesis_hash", stateSrvc.Block.GenesisHash().String()),
+		telemetry.NewKeyValue("implementation", sysSrvc.SystemName()), telemetry.NewKeyValue("msg", "system.connected"),
+		telemetry.NewKeyValue("name", cfg.Global.Name), telemetry.NewKeyValue("network_id", networkSrvc.NetworkState().PeerID),
+		telemetry.NewKeyValue("startup_time", strconv.FormatInt(time.Now().UnixNano(), 10)), telemetry.NewKeyValue("version", sysSrvc.SystemVersion())))
 
 	return node, nil
 }
